backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# PDF Processing
pdfplumber==0.10.3
pypdf==3.17.4

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Vector Database (using FAISS)
faiss-cpu==1.8.0
sentence-transformers==2.3.1

# LLM and AI
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.3
openai==1.10.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.3
